requirements.txt
Django==4.0.1
django-environ==0.9.0
psycopg2-binary==2.9.3
djangorestframework==3.14.0
social-auth-app-django==5.0.0
django-filter==22.1
drf-spectacular==0.24.2
marshmallow==3.19.0
marshmallow-dataclass==8.5.10
requests==2.28.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
factory-boy==3.2.1
Faker==15.3.4
pytest==7.2.0